pyproject.toml
[project]
name = "ciphr"
version = "0.1.3"
description = "Crawling and Ingesting Physics Research"
authors = [{ name = "Pratik Jawahar" }]
requires-python = ">=3.10, <3.13"
readme = "README.md"

dependencies = [
    "firecrawl-py",
    "google-generativeai",
    "arxiv",
    "python-dotenv",
    "pypdf",
]

[project.scripts]
ciphr = "ciphr.ciphr:main"

[dependency-groups]
dev = [
    "black>=25.1.0,<26",
    "pylint>=3.3.4,<4",
    "wrapt>=1.17.2,<2",
    "sphinx>=8.1.3,<9",
    "sphinx-rtd-theme>=3.0.2,<4",
    "pytest>=8.4.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.13",
]

[tool.ruff.lint]
extend-select = ["B"]

[tool.uv]

[tool.hatch.build.targets.sdist]
include = ["./ciphr", "README.md", "pyproject.toml"]

[tool.hatch.build.targets.wheel]
#include = ["./bead"]

[tool.hatch.build.targets.wheel.sources]
"ciphr" = "ciphr"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
